# docs/openapi_v9.yaml
openapi: 3.0.3
info:
  title: Plan Charge v9 API
  version: 1.0.0
  description: |
    OpenAPI spec for Plan Charge v9 backend (FastAPI) covering identity, planning, integrations, and reporting.
servers:
  - url: /api/v1
...
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code: { type: string }
            message: { type: string }
            details: { type: object, additionalProperties: true }
...
    BenefitType:
      type: object
      properties:
        id: { type: string, format: uuid }
        key: { type: string }
        name: { type: string }
    BenefitPolicy:
      type: object
      properties:
        id: { type: string, format: uuid }
        benefit_type_id: { type: string, format: uuid }
        rules: { type: object, additionalProperties: true }
        active: { type: boolean }
        effective_from: { type: string, format: date, nullable: true }
        effective_to: { type: string, format: date, nullable: true }
...
paths:
  /auth/login:
    post:
      tags: [Auth]
      summary: Login with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string }
      responses:
        '200':
          description: Tokens and user info
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token: { type: string }
                  refresh_token: { type: string }
                  user: { type: object }
        '401': { description: Unauthorized, content: { application/json: { schema: { $ref: '#/components/schemas/Error' }}}}
  /auth/refresh:
    post:
      tags: [Auth]
      summary: Refresh JWT tokens
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refresh_token]
              properties:
                refresh_token: { type: string }
      responses:
        '200':
          description: New tokens
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token: { type: string }
                  refresh_token: { type: string }
...
  /benefits/types:
    get:
      tags: [Benefits]
      summary: List benefit types
      responses:
        '200': { description: OK }
    post:
      tags: [Benefits]
      summary: Create benefit type
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/BenefitType' }
      responses:
        '201': { description: Created }
  /benefits/policies:
    get:
      tags: [Benefits]
      summary: List benefit policies
      responses:
        '200': { description: OK }
    post:
      tags: [Benefits]
      summary: Create benefit policy
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/BenefitPolicy' }
      responses:
        '201': { description: Created }
...